Move gogogate2 base entity to separate module
