+++ homeassistant/components/gogogate2/sensor.py
@@ -16,8 +16,9 @@ from homeassistant.const import PERCENTAGE, EntityCategory, UnitOfTemperature
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import GoGoGate2Entity, get_data_update_coordinator, sensor_unique_id
+from .common import get_data_update_coordinator, sensor_unique_id
 from .coordinator import DeviceDataUpdateCoordinator
+from .entity import GoGoGate2Entity
 
 SENSOR_ID_WIRED = "WIRE"
